Simplify streaming module: consolidate files, flatten entry points, cooperative cancel
- Merge types.py + events.py + protocol.py into types.py (all data structures in one file)
- Delete completion.py, session.py (inlined into runtime.py)
- Simplify cancellation: remove monitor task pattern, use cooperative cancel with asyncio.wait racing
- Flatten entry points: remove _run_stream and _build_instance indirection
- Single public entry point execute_chat(request, sandbox_service, session_factory)
- Fix # type: ignore in sandbox.py by narrowing type properly
- Make emit_event and redis public (accessed cross-file from context_usage.py)
