+++ backend/app/services/streaming/__init__.py
@@ -1,17 +1,18 @@
-from app.services.streaming.cancellation import CancellationHandler, StreamCancelled
+from app.services.streaming.cancellation import CancellationHandler
 from app.services.streaming.runtime import ChatStreamRuntime
-from app.services.streaming.events import ActiveToolState, StreamEvent, ToolPayload
+from app.services.streaming.types import (
+    ActiveToolState,
+    ChatStreamRequest,
+    StreamEvent,
+    ToolPayload,
+)
 from app.services.streaming.processor import StreamProcessor
-from app.services.streaming.types import ChatStreamRequest
-from app.services.streaming.session import SessionUpdateCallback
 
 __all__ = [
     "ActiveToolState",
     "ChatStreamRequest",
     "ChatStreamRuntime",
     "CancellationHandler",
-    "SessionUpdateCallback",
-    "StreamCancelled",
     "StreamEvent",
     "StreamProcessor",
     "ToolPayload",
